requirements.txt
customtkinter>=5.2.0
pyusb>=1.2.1
libusb>=1.0.27
Pillow>=10.0.0
pyinstaller>=5.13.0
